requirements.txt
python-telegram-bot
prettytable
asyncio
